+++ src/query/storages/fuse/src/table_functions/clustering_information/mod.rs
@@ -15,8 +15,6 @@
 #[allow(clippy::module_inception)]
 mod clustering_information;
 mod clustering_information_table;
-mod table_args;
 
 pub use clustering_information::ClusteringInformation;
 pub use clustering_information_table::ClusteringInformationTable;
-pub use table_args::*;
