+++ src/query/storages/fuse/src/table_functions/fuse_blocks/fuse_block_table.rs
@@ -26,13 +26,13 @@ use common_meta_app::schema::TableIdent;
 use common_meta_app::schema::TableInfo;
 use common_meta_app::schema::TableMeta;
 
-use super::table_args::parse_func_table_args;
 use crate::pipelines::processors::port::OutputPort;
 use crate::pipelines::processors::processor::ProcessorPtr;
 use crate::pipelines::processors::AsyncSource;
 use crate::pipelines::processors::AsyncSourcer;
 use crate::pipelines::Pipeline;
 use crate::sessions::TableContext;
+use crate::table_functions::parse_db_tb_ssid_args;
 use crate::table_functions::string_literal;
 use crate::table_functions::FuseBlock;
 use crate::table_functions::TableArgs;
@@ -57,7 +57,7 @@ impl FuseBlockTable {
         table_args: TableArgs,
     ) -> Result<Arc<dyn TableFunction>> {
         let (arg_database_name, arg_table_name, arg_snapshot_id) =
-            parse_func_table_args(&table_args, FUSE_FUNC_BLOCK)?;
+            parse_db_tb_ssid_args(&table_args, FUSE_FUNC_BLOCK)?;
 
         let engine = FUSE_FUNC_BLOCK.to_owned();
