+++ src/query/storages/fuse/src/table_functions/fuse_snapshots/fuse_snapshot_table.rs
@@ -30,10 +30,10 @@ use common_pipeline_sources::AsyncSource;
 use common_pipeline_sources::AsyncSourcer;
 
 use super::fuse_snapshot::FuseSnapshot;
-use super::table_args::parse_func_history_args;
 use crate::pipelines::processors::port::OutputPort;
 use crate::pipelines::Pipeline;
 use crate::sessions::TableContext;
+use crate::table_functions::parse_db_tb_args;
 use crate::table_functions::string_literal;
 use crate::table_functions::TableArgs;
 use crate::table_functions::TableFunction;
@@ -56,7 +56,7 @@ impl FuseSnapshotTable {
         table_args: TableArgs,
     ) -> Result<Arc<dyn TableFunction>> {
         let (arg_database_name, arg_table_name) =
-            parse_func_history_args(&table_args, FUSE_FUNC_SNAPSHOT)?;
+            parse_db_tb_args(&table_args, FUSE_FUNC_SNAPSHOT)?;
 
         let engine = FUSE_FUNC_SNAPSHOT.to_owned();
