+++ src/query/storages/fuse/src/table_functions/fuse_snapshots/mod.rs
@@ -14,8 +14,6 @@
 
 mod fuse_snapshot;
 mod fuse_snapshot_table;
-mod table_args;
 
 pub use fuse_snapshot::FuseSnapshot;
 pub use fuse_snapshot_table::FuseSnapshotTable;
-pub use table_args::parse_func_history_args;
