+++ src/query/storages/fuse/src/table_functions/table_args.rs
@@ -14,7 +14,10 @@
 
 use common_exception::ErrorCode;
 use common_exception::Result;
+use common_expression::Expr;
 use common_expression::Scalar;
+
+use crate::table_functions::TableArgs;
 
 pub fn string_value(value: &Scalar) -> Result<String> {
     match value {
@@ -26,4 +29,44 @@ pub fn string_value(value: &Scalar) -> Result<String> {
 
 pub fn string_literal(val: &str) -> Scalar {
     Scalar::String(val.as_bytes().to_vec())
+}
+
+pub fn parse_db_tb_args(table_args: &TableArgs, func_name: &str) -> Result<(String, String)> {
+    let args = table_args.expect_all_positioned(func_name, Some(2))?;
+    let db = string_value(&args[0])?;
+    let tbl = string_value(&args[1])?;
+    Ok((db, tbl))
+}
+
+pub fn parse_db_tb_ssid_args(
+    table_args: &TableArgs,
+    func_name: &str,
+) -> Result<(String, String, Option<String>)> {
+    let args = table_args.expect_all_positioned(func_name, None)?;
+    match args.len() {
+        3 => {
+            let db = string_value(&args[0])?;
+            let tbl = string_value(&args[1])?;
+            let snapshot_id = string_value(&args[2])?;
+            Ok((db, tbl, Some(snapshot_id)))
+        }
+        2 => {
+            let db = string_value(&args[0])?;
+            let tbl = string_value(&args[1])?;
+            Ok((db, tbl, None))
+        }
+        _ => Err(ErrorCode::BadArguments(format!(
+            "expecting <database>, <table_name> and <snapshot_id> (as string literals), but got {:?}",
+            args
+        ))),
+    }
+}
+
+pub fn unwrap_tuple(expr: &Expr) -> Option<Vec<Expr>> {
+    match expr {
+        Expr::FunctionCall { function, args, .. } if function.signature.name == "tuple" => {
+            Some(args.clone())
+        }
+        _ => None,
+    }
 }
